Create Connect_Exchange_Online_MFA.ps1
I quickly whipped this up because the recommended way of connecting to Exchange Online with MFA enabled is laborious/annoying and the documentation doesn't tell you that you have to use Internet Explorer. It works for me. If it doesn't work for you try this one that I found after making this: https://www.powershellgallery.com/packages/Load-ExchangeMFA/1.1/Content/Load-ExchangeMFA.ps1

https://docs.microsoft.com/en-us/powershell/exchange/exchange-online/connect-to-exchange-online-powershell/mfa-connect-to-exchange-online-powershell?view=exchange-ps